+++ docs/nuxt.config.js
@@ -1,14 +1,21 @@
+import { resolve } from 'path'
 import { withDocus } from 'docus'
+import windiConfig from './windi.config'
 
 export default withDocus({
   content: {
     liveEdit: false
   },
   buildModules: [
-    'vue-plausible/lib/esm'
+    'vue-plausible'
   ],
   plausible: {
     // https://github.com/moritzsternemann/vue-plausible#configuration
     domain: 'docus.dev'
+  },
+  windicss: {
+    root: resolve(__dirname, '..'),
+    // Workaround waiting for https://github.com/windicss/nuxt-windicss-module/issues/30
+    config: windiConfig
   }
 })