+++ theme/components/molecules/DColorSwitcher.vue
@@ -2,7 +2,7 @@
   <span class="inline-flex">
     <button
       slot="placeholder"
-      class="w-6 h-6 text-gray-500 transition-colors duration-200 dark:text-gray-500 hover:text-gray-500 dark:hover:text-gray-400 focus:outline-none"
+      class="w-6 h-6 text-gray-500 transition-colors duration-200 dark:text-gray-500 hover:text-gray-700 dark:hover:text-gray-400 focus:outline-none"
       aria-label="Color Mode"
       @click="$colorMode.value === 'dark' ? $colorMode.preference = 'light' : $colorMode.preference = 'dark'"
     >