+++ theme/nuxt.config.js
@@ -40,8 +40,8 @@ export default {
   ],
   buildModules: [
     themeModule,
+    'nuxt-windicss',
     'nuxt-vite',
-    '@nuxtjs/tailwindcss',
     '@nuxtjs/color-mode',
     '@nuxtjs/pwa',
     '@nuxtjs/google-fonts'
@@ -64,7 +64,7 @@ export default {
   content: {
     markdown: {
       prism: {
-        theme: ''
+        theme: r('assets/css/prism.css')
       },
       remarkPlugins: [
         [r('utils/remark-prose'), {
@@ -119,9 +119,6 @@ export default {
       Inter: true
     }
   },
-  tailwindcss: {
-    jit: true
-  },
   server: {
     port: process.env.PORT || 4000
   },
@@ -131,6 +128,11 @@ export default {
         'ohmyfetch',
         'vue-demi',
         '@vueuse/integrations'
+      ],
+      include: [
+        'defu',
+        'theme-colors',
+        'lodash.groupby'
       ]
     },
     vue: {}